Add status button when editing inside Working Copy location.
Using a file provider service the button will show the number of lines added or deleted compared to last commit.
Tap the button to start a commit inside Working Copy that will jump back into Textor when done or cancelled.

These changes only appear for files inside Working Copy and with a recent version of the app installed and takes care not to use resources when there is no file provider service to use.